requirements-dev.txt
configparser==3.5.0
coverage==4.2
enum34==1.1.6
flake8==3.2.0
mccabe==0.5.2
Pillow==3.4.2
py==1.4.31
pycodestyle==2.2.0
pyflakes==1.3.0
pytest==3.0.4
